+++ cmd/zc_traverser_benchmark.go
@@ -47,6 +47,18 @@ func (t *benchmarkTraverser) isDirectory(bool) bool {
 	return true
 }
 
+func (_ *benchmarkTraverser) toReversedString(i uint) string {
+	s := fmt.Sprintf("%d", i)
+	count := len(s)
+	b := []byte(s)
+	r := make([]byte, count)
+	lastIndex := count - 1
+	for n, x := range b {
+		r[lastIndex-n] = x
+	}
+	return string(r)
+}
+
 func (t *benchmarkTraverser) traverse(preprocessor objectMorpher, processor objectProcessor, filters []objectFilter) (err error) {
 	if len(filters) > 0 {
 		panic("filters not expected or supported in benchmark traverser") // but we still call processIfPassedFilters below, for consistency with other traversers
@@ -54,7 +66,7 @@ func (t *benchmarkTraverser) traverse(preprocessor objectMorpher, processor obje
 
 	for i := uint(1); i <= t.fileCount; i++ {
 
-		name := fmt.Sprintf("%d", i)
+		name := t.toReversedString(i) // this gives an even distribution through the namespace (compare the starting characters, for 0 to 199, when reversed or not). This is useful for performance when High Throughput Block Blob pathway does not apply
 		relativePath := name
 
 		if t.incrementEnumerationCounter != nil {