+++ testSuite/scripts/run.py
@@ -16,6 +16,7 @@ from test_blobfs_upload_SAS import *
 from test_autodetect_blob_type import *
 from test_file_sync import *
 from test_file_copy import *
+from test_clfsload import *
 import glob, os
 import configparser
 import platform
@@ -207,7 +208,8 @@ def main():
     print("Smoke tests starting...")
     init()
 
-    test_class_to_run = [BlobPipingTests,
+    test_class_to_run = [LoadUserScenario,
+                         BlobPipingTests,
                          File_Service_2_Service_Copy_User_Scenario,
                          Blob_Sync_User_Scenario,
                          File_Sync_User_Scenario,